gpt_sql_webapp/requirements.txt
openai
gradio
python-dotenv
langchain
langchain-openai
langchain-community
psycopg2-binary
sqlalchemy
